requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
google-genai==1.69.0
python-dotenv==1.2.2
pydantic==2.10.1
pydantic[email]